Refactor BigQueryResponseJsonProtocol to client.ResponseJsonProtocol
