+++ google-cloud-bigquery/src/test/java/docs/javadsl/GoogleBigQuerySourceDoc.java
@@ -5,6 +5,7 @@
 package docs.javadsl;
 
 // #imports
+
 import akka.NotUsed;
 import akka.actor.ActorSystem;
 import akka.http.scaladsl.model.HttpRequest;
